EY-2646: hentVedtak-biten er også ganske så uavhengig, så trekkjer også den ut til si eiga service. Prøver å nærme oss prinsippet om ei klasse, eitt ansvar
