+++ apps/etterlatte-brev-api/src/main/kotlin/no/nav/etterlatte/brev/VedtaksbrevService.kt
@@ -8,6 +8,7 @@ import no.nav.etterlatte.brev.brevbaker.BrevbakerService
 import no.nav.etterlatte.brev.db.BrevRepository
 import no.nav.etterlatte.brev.dokarkiv.DokarkivServiceImpl
 import no.nav.etterlatte.brev.hentinformasjon.ISakOgBehandlingService
+import no.nav.etterlatte.brev.hentinformasjon.VedtaksvurderingService
 import no.nav.etterlatte.brev.journalpost.JournalpostResponse
 import no.nav.etterlatte.brev.model.Brev
 import no.nav.etterlatte.brev.model.BrevData
@@ -34,6 +35,7 @@ import java.util.UUID
 class VedtaksbrevService(
     private val db: BrevRepository,
     private val sakOgBehandlingService: ISakOgBehandlingService,
+    private val vedtaksvurderingService: VedtaksvurderingService,
     private val adresseService: AdresseService,
     private val dokarkivService: DokarkivServiceImpl,
     private val brevbaker: BrevbakerService,
@@ -127,7 +129,7 @@ class VedtaksbrevService(
         }
 
         val (saksbehandlerIdent, vedtakStatus) =
-            sakOgBehandlingService.hentVedtakSaksbehandlerOgStatus(
+            vedtaksvurderingService.hentVedtakSaksbehandlerOgStatus(
                 brev.behandlingId!!,
                 brukerTokenInfo,
             )
